+++ lib/index.d.ts
@@ -150,6 +150,11 @@ declare namespace Realm {
          */
         isValid(): boolean;
 
+        /**
+         * @returns boolean
+         */
+        isEmpty(): boolean;
+
         min(property?: string): number | Date | null;
         max(property?: string): number | Date | null;
         sum(property?: string): number | null;
@@ -669,6 +674,7 @@ declare class Realm {
      * @returns void
      */
     addListener(name: string, callback: (sender: Realm, event: 'change') => void): void;
+    addListener(name: string, callback: (sender: Realm, event: 'schema', schema: Realm.ObjectSchema[]) => void): void;
 
     /**
      * @param  {string} name
@@ -676,6 +682,7 @@ declare class Realm {
      * @returns void
      */
     removeListener(name: string, callback: (sender: Realm, event: 'change') => void): void;
+    removeListener(name: string, callback: (sender: Realm, event: 'schema', schema: Realm.ObjectSchema[]) => void): void;
 
     /**
      * @param  {string} name?