+++ tests/js/realm-tests.js
@@ -803,7 +803,7 @@ module.exports = {
         TestCase.assertEqual(secondNotificationCount, 1);
 
         TestCase.assertThrowsContaining(() => realm.addListener('invalid', () => {}),
-                                        "Only the 'change' notification name is supported.");
+                                        "Only the 'change' and 'schema' notification names are supported.");
 
         realm.addListener('change', () => {
             throw new Error('expected error message');
@@ -1219,6 +1219,43 @@ module.exports = {
             new Realm({ path: 'dates-v3.realm', disableFormatUpgrade: true } );
         }, 'The Realm file format must be allowed to be upgraded in order to proceed.');
     },
+
+
+    // FIXME: We need to test adding a property also calls the listener
+    testSchemaUpdatesNewClass: function() {
+        return new Promise((resolve, reject) => {
+            let realm1 = new Realm({ _cache: false });
+            TestCase.assertTrue(realm1.empty);
+            TestCase.assertEqual(realm1.schema.length, 0);  // empty schema
+            realm1.addListener('schema', (realm, event, schema) => {
+                TestCase.assertEqual(event, 'schema');
+                TestCase.assertEqual(schema.length, 1);
+                TestCase.assertEqual(realm.schema.length, 1);
+                TestCase.assertEqual(schema[0].name, 'TestObject');
+                TestCase.assertEqual(realm1.schema.length, 1);
+                TestCase.assertEqual(realm.schema[0].name, 'TestObject');
+            });
+
+            const schema = [{
+                name: 'TestObject',
+                properties: {
+                    prop0: 'string',
+                }
+            }];
+
+            let realm2 = new Realm({ schema: schema, _cache: false });
+            TestCase.assertEqual(realm1.schema.length, 0); // not yet updated
+            TestCase.assertEqual(realm2.schema.length, 1);
+
+            // give some time to let advance_read to complete
+            // in real world, a Realm will not be closed just after its
+            // schema has been updated
+            setTimeout(() => {
+                resolve();
+            }, 1000);
+        });
+    },
+
 
     // FIXME: reanble test
     /*