+++ CHANGELOG.md
@@ -1,18 +1,47 @@
-x.x.x Release notes (xxxx-xx-xx)
+X.Y.Z Release notes
 =============================================================
 ### Compatibility
+* Sync protocol: 24
+* Server-side history format: 4
+* File format: 7
+* Realm Object Server: 3.0.0 or later
 
 ### Breaking changes
 * None.
 
 ### Enhancements
-* None
+* None.
 
 ### Bug fixes
-* Fixed `Realm.open()` to work without passing a config.
 * Fixed TypeScript definition of `Realm.objects()` and `Realm.objectForPrimaryKety()`.
 
 ### Internal
+* Realm Sync v3.3.0.
+* Realm Core v5.6.0.
+
+
+2.7.0 Release notes (2018-5-29)
+=============================================================
+### Compatibility
+* Sync protocol: 24
+* Server-side history format: 4
+* File format: 7
+* Realm Object Server: 3.0.0 or later
+
+### Breaking changes
+* None.
+
+### Enhancements
+* Added `isEmpty()` method on `Realm.Results` and `Realm.List`.
+* Added schema change listener to `Realm.addListener()` (#1825).
+
+### Bug fixes
+* Fixed `Realm.open()` to work without passing a config.
+* Fixed a bug in `Realm.open()` to work without passing a config.
+
+### Internal
+* Realm Sync v3.3.0.
+* Realm Core v5.6.0.
 
 
 2.6.0 Release notes (2018-5-16)